requirements.txt
requests
PyYAML
Jinja2
python-dateutil
tenacity
